cmdargs.py script added as target (another exe) to build_win

+++ build/build_win.py
@@ -13,12 +13,19 @@ import subprocess
 import platform
 import shutil
 
-os.chdir("..")
-print "Working directory set to: ", os.getcwd()
 
 VERSION = "0.7.0"                   # todo: get version and revision from GIT
+ZIP_ENABLED = False
+
 PRJ_SPEC_FILE = "woofer.spec"
 PYINSTALLER_EXE_NAME = "pyinstaller"
+
+
+build_dir = os.path.dirname(os.path.realpath(sys.argv[0]))
+root_dir = os.path.dirname(build_dir)
+os.chdir(root_dir)
+print "Root directory set to: ", os.path.abspath(root_dir)
+print "Working directory set to: ", os.getcwd()
 
 
 def which(program):
@@ -53,8 +60,8 @@ def main():
     else:
         print "PyInstaller executable found at:", pyinstaller_exe
 
-    spec_file = os.path.join(os.getcwd(), "build", PRJ_SPEC_FILE)
-    dist_path = os.path.join(os.getcwd(), "dist", "woofer")
+    spec_file = os.path.join(root_dir, "build", PRJ_SPEC_FILE)
+    dist_path = os.path.join(root_dir, "dist", "woofer")
     if not os.path.isfile(spec_file):
         raise Exception("Unable to locate spec file at: %s" % spec_file)
     else:
@@ -70,7 +77,7 @@ def main():
     if retcode != 0:
         raise Exception("Building finished with error code: %s!!!" % retcode)
 
-    new_dist_path = os.path.join(os.getcwd(), "build", "release", "woofer_%s_v%s"
+    new_dist_path = os.path.join(root_dir, "build", "release", "woofer_%s_v%s"
                                  % (platform.architecture()[0], VERSION))
     if os.path.isdir(new_dist_path):
         shutil.rmtree(new_dist_path)
@@ -80,16 +87,17 @@ def main():
 
     print "Dist package successfully built to:", dist_path
 
-    decision = raw_input("Are you want to ZIP built dist folder? (y/n) ")
-    if decision != 'y':
-        return
+    if ZIP_ENABLED:
+        decision = raw_input("Are you want to ZIP built dist folder? (y/n) ")
+        if decision != 'y':
+            return
 
-    zip_folder = shutil.make_archive(dist_path, "zip", dist_path)
-    print "Dist zipped successfully to:", zip_folder
+        zip_folder = shutil.make_archive(dist_path, "zip", dist_path)
+        print "Dist zipped successfully to:", zip_folder
 
-    print "Removing dist folder..."
-    shutil.rmtree(dist_path)
-    shutil.rmtree(os.path.join(os.getcwd(), "dist"))
+        print "Removing dist folder..."
+        shutil.rmtree(dist_path)
+        shutil.rmtree(os.path.join(root_dir, "dist"))
 
 
 if __name__ == "__main__":
